requirements.txt
streamlit
torch
torchvision
torchaudio
pillow
numpy
